requirements.txt
mpi4py
web3
py-solc-x
opencv-python
